+++ configure.py
@@ -923,6 +923,7 @@ scylla_core = (['message/messaging_service.cc',
                 'db/consistency_level.cc',
                 'db/system_keyspace.cc',
                 'db/virtual_table.cc',
+                'db/virtual_tables.cc',
                 'db/system_distributed_keyspace.cc',
                 'db/size_estimates_virtual_reader.cc',
                 'db/schema_tables.cc',
